adjust gem post install hook to allow updating the default gem
rubygems/defaults/jruby.rb needs to _not_ require a file from
jar-dependencies when it first runs, otherwise bundler can't
load a version of jar-dependencies that's installed later.
this commit is intended to have jruby.rb register the
post-install hook itself, and therefore defer loading
jar-dependencies until it's actually needed. jruby needs
an update for its defaults to match the new
jar_install_post_install_hook.rb. in the meantime, this
commit is compatible with the current defaults or an
updated one.